+++ wechatpy/pay/__init__.py
@@ -42,8 +42,6 @@ class WeChatPay(object):
     """红包接口"""
     transfer = api.WeChatTransfer()
     """企业付款接口"""
-    transfer_bankcard = api.WeChatTransferBankCard()
-    """企业付款到银行卡接口"""
     coupon = api.WeChatCoupon()
     """代金券接口"""
     order = api.WeChatOrder()
