+++ wechatpy/pay/api/__init__.py
@@ -2,7 +2,6 @@
 from __future__ import absolute_import, unicode_literals
 from wechatpy.pay.api.redpack import WeChatRedpack  # NOQA
 from wechatpy.pay.api.transfer import WeChatTransfer  # NOQA
-from wechatpy.pay.api.transfer_bankcard import WeChatTransferBankCard  # NOQA
 from wechatpy.pay.api.coupon import WeChatCoupon  # NOQA
 from wechatpy.pay.api.order import WeChatOrder  # NOQA
 from wechatpy.pay.api.refund import WeChatRefund  # NOQA
