+++ wechatpy/pay/utils.py
@@ -7,6 +7,11 @@ import hashlib
 import socket
 
 import six
+from cryptography.hazmat.backends import default_backend
+from cryptography.hazmat.primitives import serialization
+from cryptography.hazmat.primitives import hashes
+from cryptography.hazmat.primitives.asymmetric import padding
+
 
 from wechatpy.utils import to_binary, to_text
 
@@ -56,36 +61,46 @@ def get_external_ip():
         return '127.0.0.1'
 
 
-def rsa_encrypt(data, pem):
+def rsa_encrypt(data, pem, b64_encode=True):
     """
-    加密
+    rsa 加密
     :param data: 待加密字符串
-    :param pem: RSA key 内容
+    :param pem: RSA public key 内容
+    :param b64_encode: 是否对输出进行 base64 encode
     :return:
     """
-    encoded_data = data.encode() if not isinstance(data, bytes) else data
-    try:
-        from Crypto.PublicKey import RSA
-        from Crypto.Cipher import PKCS1_OAEP
-        rsakey = RSA.importKey(pem)
-        cipher = PKCS1_OAEP.new(rsakey)
-        encrypted_data = cipher.encrypt(encoded_data)
-    except ModuleNotFoundError:
-        from cryptography.hazmat.backends import default_backend
-        from cryptography.hazmat.primitives import serialization
-        from cryptography.hazmat.primitives import hashes
-        from cryptography.hazmat.primitives.asymmetric import padding
-
-        private_key = serialization.load_pem_private_key(pem, password=None, backend=default_backend())
-        encrypted_data = private_key.sign(
-            data=encoded_data,
-            padding=padding.OAEP(
-                mgf=padding.MGF1(hashes.SHA1()),
-                algorithm=hashes.SHA1(),
-                label=None,
-            ),
-            algorithm=hashes.SHA1()
+    encoded_data = data.encode('utf-8') if not isinstance(data, bytes) else data
+    pem = pem.encode('utf-8') if not isinstance(pem, bytes) else pem
+    public_key = serialization.load_pem_public_key(pem, backend=default_backend())
+    encrypted_data = public_key.encrypt(
+        encoded_data,
+        padding=padding.OAEP(
+            mgf=padding.MGF1(hashes.SHA1()),
+            algorithm=hashes.SHA1(),
+            label=b'',
         )
-    except ModuleNotFoundError:
-        raise ModuleNotFoundError('either crypto or cryptography is required')
-    return base64.b64encode(encrypted_data).decode()
+    )
+    if b64_encode:
+        encrypted_data = base64.b64encode(encrypted_data).decode('utf-8')
+    return encrypted_data
+
+
+def rsa_decrypt(encrypted_data, pem, password=None):
+    """
+    rsa 解密
+    :param encrypted_data: 待解密 bytes
+    :param pem: RSA private key 内容
+    :param password: RSA private key pass phrase
+    :return:
+    """
+    pem = pem.encode('utf-8') if not isinstance(pem, bytes) else pem
+    private_key = serialization.load_pem_private_key(pem, password, backend=default_backend())
+    data = private_key.decrypt(
+        encrypted_data,
+        padding=padding.OAEP(
+            mgf=padding.MGF1(hashes.SHA1()),
+            algorithm=hashes.SHA1(),
+            label=b'',
+        )
+    )
+    return data
